Fix issue with ordereddict.OrderedDict
The constructor of ordereddict.OrderedDict calls the `clear` method
which is responsible for creating a member variable. Since in Dicti
this method is overwritten the member never got created.
To fix this a different (and more efficient) implementation of `clear`
is now used that makes use of the clear method of the base type.