Enable test file execution from the root folder
Individual test files can now be executed as follows:

- python -m test.test_pydicti
- python test/test_pydicti.py

The first usage requires a '__init__.py' file in the test folder.
Furthermore on python3 the import of test_compat needs to be either
absolute or relative.

For the second case a relative import can't be used. Therefore an
absolute import must be used.